requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
build==0.10.0
CacheControl==0.13.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
contourpy==1.2.1
crashtest==0.4.1
cryptography==42.0.7
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
distlib==0.3.8
docker-pycreds==0.4.0
dulwich==0.21.7
evaluate==0.4.2
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.17.3
keyring==24.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.0
pandas==2.2.2
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.10.0
platformdirs==3.11.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.6.0
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.11.2
pyparsing==3.1.2
pyproject_hooks==1.1.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
rapidfuzz==3.9.1
regex==2024.5.15
requests==2.32.2
requests-toolbelt==1.0.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
SecretStorage==3.3.3
sentencepiece==0.2.0
sentry-sdk==2.7.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sympy==1.12.1
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.5
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
trove-classifiers==2024.5.22
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
wandb==0.17.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.0
